+++ lib/osm/event.rb
@@ -311,11 +311,6 @@ module Osm
 
     private
     def self.new_event_from_data(event_data)
-      columns = []
-      ActiveSupport::JSON.decode(event_data['config']).each do |field|
-        columns.push Column.new(:id => field['id'], :name => field['name'], :parent_label => field['pL'])
-      end
-
       event = Osm::Event.new(
         :id => Osm::to_i_or_nil(event_data['eventid']),
         :section_id => Osm::to_i_or_nil(event_data['sectionid']),
@@ -326,13 +321,19 @@ module Osm
         :location => event_data['location'],
         :notes => event_data['notes'],
         :archived => event_data['archived'].eql?('1'),
-        :columns => columns,
         :notepad => event_data['notepad'],
         :public_notepad => event_data['publicnotes'],
         :confirm_by_date => Osm::parse_date(event_data['confdate']),
         :allow_changes => event_data['allowchanges'].eql?('1'),
         :reminders => !event_data['disablereminders'].eql?('1'),
       )
+
+      columns = []
+      ActiveSupport::JSON.decode(event_data['config']).each do |field|
+        columns.push Column.new(:id => field['id'], :name => field['name'], :label => field['pL'], :event => event)
+      end
+      event.columns = columns
+      return event
     end
 
 
@@ -344,14 +345,17 @@ module Osm
       #   @return [String] OSM id for the column
       # @!attribute [rw] name
       #   @return [String] name for the column (displayed in OSM)
-      # @!attribute [rw] parent_label
+      # @!attribute [rw] label
       #   @return [String] label to display in My.SCOUT ("" prevents display in My.SCOUT)
+      # @!attriute [rw] event
+      #   @return [Osm::Event] the event that this column belongs to
 
       attribute :id, :type => String
       attribute :name, :type => String
-      attribute :parent_label, :type => String, :default => ''
+      attribute :label, :type => String, :default => ''
+      attribute :event
 
-      attr_accessible :id, :name, :parent_label
+      attr_accessible :id, :name, :label, :event
 
       validates_presence_of :id
       validates_presence_of :name
@@ -360,6 +364,28 @@ module Osm
       # @!method initialize
       #   Initialize a new Column
       #   @param [Hash] attributes the hash of attributes (see attributes for descriptions, use Symbol of attribute name as the key)
+
+
+      # Update event column
+      # @param [Osm::Api] api The api to use to make the request
+      # @return [Boolean] if the operation suceeded or not
+      def update(api)
+        raise Forbidden, 'you do not have permission to write to events for this section' unless Osm::Model.get_user_permission(api, event.section_id, :events).include?(:write)
+
+        data = api.perform_query("events.php?action=renameColumn&sectionid=#{event.section_id}&eventid=#{event.id}", {
+          'columnId' => id,
+          'columnName' => name,
+          'pL' => label
+        })
+
+        (ActiveSupport::JSON.decode(data['config']) || []).each do |i|
+          if i['id'] == id
+            return i['name'].eql?(name) && (i['pL'].nil? || i['pL'].eql?(label))
+          end
+        end
+        return false
+
+      end
 
     end # class Column
 
@@ -376,6 +402,8 @@ module Osm
       #   @return [Hash] Keys are the field's id, values are the field values
       # @!attribute [rw] row
       #   @return [Fixnum] part of the OSM API
+      # @!attriute [rw] event
+      #   @return [Osm::Event] the event that this attendance applies to
   
       attribute :row, :type => Integer
       attribute :member_id, :type => Integer